Added version parsing to MATCH arms
Now that we don't use the API to retrieve the study arms any more, there
is need to try to get the latest version of each (this was done by a API
arg previously). Added a method to get the versions of each arm
iteration, and create a dict `self._latest_ver` to track the latest arm
versions for the sake of mapping.